+++ src/main/java/org/opensearch/knn/index/warmup/MemoryOptimizedSearchWarmup.java
@@ -19,12 +19,12 @@ import org.apache.lucene.store.IndexInput;
 import org.opensearch.common.lucene.Lucene;
 import org.opensearch.index.mapper.MappedFieldType;
 import org.opensearch.index.mapper.MapperService;
-import org.opensearch.knn.index.VectorDataType;
 import org.opensearch.knn.index.codec.util.KNNCodecUtil;
 import org.opensearch.knn.index.engine.KNNEngine;
 import org.opensearch.knn.index.engine.MemoryOptimizedSearchSupportSpec;
 import org.opensearch.knn.index.mapper.KNNVectorFieldMapper;
 import org.opensearch.knn.index.mapper.KNNVectorFieldType;
+import org.opensearch.knn.memoryoptsearch.faiss.FaissMemoryOptimizedSearcher;
 
 import java.io.IOException;
 import java.nio.file.Path;
@@ -33,7 +33,6 @@ import java.util.ArrayList;
 import java.util.List;
 
 import static org.opensearch.knn.common.FieldInfoExtractor.extractKNNEngine;
-import static org.opensearch.knn.common.KNNConstants.VECTOR_DATA_TYPE_FIELD;
 
 @Log4j2
 public class MemoryOptimizedSearchWarmup {
@@ -96,18 +95,10 @@ public class MemoryOptimizedSearchWarmup {
         }
 
         try {
-            final String dataTypeStr = field.getAttribute(VECTOR_DATA_TYPE_FIELD);
-            if (dataTypeStr == null) {
-                return true;
-            }
-            final VectorDataType vectorDataType = VectorDataType.get(dataTypeStr);
-            // We expect a NPE here:
-            if (vectorDataType == VectorDataType.FLOAT) {
-                segmentReader.getVectorReader().search(field.getName(), (float[]) null, null, null);
-            } else {
-                segmentReader.getVectorReader().search(field.getName(), (byte[]) null, null, null);
-            }
-        } catch (NullPointerException ignored) {}
+            segmentReader.getVectorReader().search(field.getName(), (float[]) null, null, null); // codecov[ignore]
+        } catch (FaissMemoryOptimizedSearcher.WarmupInitializationException ignored) {
+            // Expected during warmup initialization
+        }
 
         return true;
     }