+++ CHANGELOG.md
@@ -7,3 +7,4 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ## [Unreleased 3.2](https://github.com/opensearch-project/k-NN/compare/main...HEAD)
 * Memory optimized search warmup ([#2954](https://github.com/opensearch-project/k-NN/pull/2954))
 * Changed warmup seek to use long instead of int to avoid overflow [#3067](https://github.com/opensearch-project/k-NN/pull/3067)
+* Fix memory-optimized-search reentrant search bug in byte index. [#3071](https://github.com/opensearch-project/k-NN/pull/3071)